requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
click==8.1.8
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.1
